don't load plugins lacking the required executables
some plugins need fundamental executables to work properly. e.g.:
'vim-gitgutter' does not work when 'git' is not available.

in some environments, where 'git' (or 'ctags') is not installed,
launching (neo)vim yields a warning ('xyz is not installed').

this commit stops loading plugins which are in no condition to be
loaded.