Optimize loading plugins
When Vim8/NeoVim are used, `pathogen` is not really needed to load
the files in `pack`. But when Vim8/NeoVim is loading the plugins,
this will happen AFTER vimrc/init.vim is parsed. This means, any
decision based upon a plugin is loaded or not must be postponed to when
the Vim8/NeoVim is done with loading. Once Vim is done, it will fire
`VimEnter` and this is now used to do any plugin specific setup.

Since Vim8/NeoVim are picking up the plugins, there was essentially no
way `g:pathogen_disabled` was of any use anymore. It is gone.

One some systems, `git` is not available. The only plugin confused by
this is `gitgutter`. Thus, I first set `g:gitgutter_git_executable` to
something stupid which makes the plugin be quiet. Once all plugins are
loaded, the result of the `git`-checks enables or disables `gitgutter`.